Refactor plugin subscription logic and add hooks
Extracted intra- and inter-plugin subscription logic into new hooks (UseIntraSubscription, UseInterSubscription) for better modularity and reuse. Refactored PluginWrapper to use these hooks and updated the interface for wrapped components. Moved plugin addition logic to a separate AddPlugin utility and updated PluginAdder accordingly. Cleaned up Renderer and removed unused networkAddress prop.
